+++ tests/testthat/test_applyChangeMeta.R
@@ -26,13 +26,18 @@ test_that("Changes to GADSdat on variable level", {
 })
 
 test_that("Changes to GADSdat on value level", {
-  changes_val2 <- changes_val
+  changes_val3 <- changes_val2 <- changes_val
   changes_val2[1, "valLabel_new"] <- "new_miss"
   changes_val2[2, "valLabel_new"] <- "new_miss2"
   g2 <- applyChangeMeta(changes_val2, dfSAV)
   expect_equal(g2$labels[, -7], dfSAV$labels[, -7])
   expect_equal(g2$labels$valLabel, c("new_miss", "new_miss2", "One", "missing", NA, "missing", NA))
   expect_equal(names(g2$dat), names(dfSAV$dat))
+
+  changes_val3[3, "missings_new"] <- "miss"
+  g3 <- applyChangeMeta(changes_val3, dfSAV)
+  expect_equal(g3$labels[, -8], dfSAV$labels[, -8])
+  expect_equal(g3$labels$missings[1:3], rep("miss", 3))
 })
 
 test_that("Changes to GADSdat: recoding", {
@@ -51,23 +56,74 @@ test_that("recoding if potential danger of overwriting old values!", {
   chang2 <- chang
   chang2[, "value_new"] <- c(3, 4, 1, NA, NA, NA)
   out <-  applyChangeMeta(chang2, df_rec)
-  expect_equal(out$labels$value, c(3, 4, 1, 1, 2, 3))
+  expect_equal(out$labels$value, c(1, 3, 4, 1, 2, 3))
+  expect_equal(out$labels$valLabel, c("z", "x", "y", "a", "b", "d"))
   expect_equal(out$dat$v1, c(3, 4, 1))
   ## partial recoding, multiple variables
-  chang[, "value_new"] <- c(3, NA, 1, 2, 1, 2)
+  chang[, "value_new"] <- c(3, NA, 1, 2, 1, 5)
   out <-  applyChangeMeta(chang, df_rec)
-  expect_equal(out$labels$value, c(3, 2, 1, 2, 1, 2))
+  expect_equal(out$labels$value, c(1, 2, 3, 1, 2, 5))
+  expect_equal(out$labels$valLabel, c("z", "y", "x", "b", "a", "d"))
   expect_equal(out$dat$v1, 3:1)
-  expect_equal(out$dat$b, c(1, 2, 2))
+  expect_equal(out$dat$b, c(1, 2, 5))
+
+  changes_val2 <- changes_val
+  changes_val2[1:2, "value_new"] <- c(-96, -95)
+  out3 <- applyChangeMeta(changes_val2, dfSAV)
+  expect_equal(out3$labels$value[1:3], c(-96, -95, 1))
+})
+
+test_that("Recoding with value meta data conflicts", {
+  changes_val2 <- changes_val
+  changes_val2[1, "value_new"] <- 1
+  expect_error(out <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "stop"),
+               "Values in 'value_new' with existing meta data in variable VAR1: 1")
+  out <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "value")
+  comp1 <- dfSAV$labels[-3, ]
+  comp1[1, "value"] <- 1
+  comp1 <- comp1[c(2, 1, 3:6),]
+  rownames(comp1) <- NULL
+  expect_equal(comp1, out)
+
+  out2 <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "value_new")
+  comp2 <- dfSAV$labels[-3, ]
+  comp2[1, "value"] <- 1
+  comp2[1, "valLabel"] <- "One"
+  comp2[1, "missings"] <- "valid"
+  comp2 <- comp2[c(2, 1, 3:6),]
+  rownames(comp2) <- NULL
+  expect_equal(comp2, out2)
+})
+
+test_that("Recoding multiple value into the same value (with meta data conflicts)", {
+  changes_val2 <- changes_val
+  changes_val2[1:2, "value_new"] <- 1
+  expect_error(out <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "stop"),
+               "Values in 'value_new' with existing meta data in variable VAR1: 1")
+  expect_error(out <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "value"),
+               "Multiple values are recoded into 1 for variable VAR1. Value meta data can thus not be used from 'value'. Set 'existingMeta' to 'value_new'.")
+
+  out2 <- recode_labels(dfSAV$labels, changes_val2, existingMeta = "value_new")
+  comp2 <- dfSAV$labels[-(2:3), ]
+  comp2[1, "value"] <- 1
+  comp2[1, "valLabel"] <- "One"
+  comp2[1, "missings"] <- "valid"
+  rownames(comp2) <- NULL
+  expect_equal(comp2, out2)
+
+  out3 <- applyChangeMeta(changes_val2, GADSdat = dfSAV, existingMeta = "value_new")
+  expect_equal(out3$dat$VAR1, c(1, 1, 1, 2))
+  expect_equal(comp2, out3$labels)
 })
 
 
 changes_val2 <- rbind(changes_val, data.frame(varName = "VAR1", value = NA, valLabel = NA, missings = NA, value_new = 2, valLabel_new = "Two", missings_new = "valid", stringsAsFactors = FALSE))
+changes_val3 <- rbind(changes_val2, data.frame(varName = "VAR1", value = NA, valLabel = NA, missings = NA, value_new = 3, valLabel_new = "Three", missings_new = "valid", stringsAsFactors = FALSE))
 
 test_that("Expand labels", {
   out <- expand_labels(df1$labels, new_varName_vec = c("ID1", "ID1", "V1"))
   expect_equal(out$varName, c("ID1", "ID1", "V1"))
-  expect_equal(out$labeled, c("yes", "yes", "no"))
+  expect_equal(out$labeled, rep("no", 3))
   out2 <- expand_labels(df2$labels, new_varName_vec = c("ID1", "V2", "V2"))
   expect_equal(out2$varName, c("ID1", "V2", "V2"))
   expect_equal(out2$labeled, c("no", "yes", "yes"))
@@ -83,10 +139,19 @@ test_that("Adding value labels for values without labels", {
   expect_equal(out$value[1:4], c(-99, -96, 1, 2))
   expect_equal(dim(out), c(8, 8))
 
-  applyChangeMeta(changes_val2, dfSAV)
-  # to do
-  # checken, dass keine Probleme in recode_dat auftauchen
-  # loest das nicht teilweise errors aus, da nicht kompatibel mit altem labels-df? wenn nein, wieso nicht?
+  df1_changes <- getChangeMeta(df1, level = "value")
+  df1_changes[2, "value_new"] <- 99
+  df1_changes[2, "valLabel_new"] <- "test"
+  out2 <- recode_labels(df1$labels, df1_changes)
+  expect_equal(out2$value, c(NA, 99))
+  expect_equal(dim(out2), c(2, 8))
+  expect_equal(out2$labeled, c("no", "yes"))
+
+  # multiple new value labels
+  out3 <- applyChangeMeta(changes_val3, dfSAV)
+  expect_equal(dim(out3$labels), c(9, 8))
+  expect_equal(out3$labels$value[3:5], 1:3)
+  expect_equal(out3$labels$valLabel[3:5], c("One", "Two", "Three"))
 })
 
 test_that("update labeled helper", {
@@ -138,7 +203,7 @@ test_that("Changes to all_GADSdat on value level", {
 })
 
 
-test_that("Changes to GADSdat if tibble, varNames", {
+test_that("Changes to GADSdat if tibble or data.frame, varNames", {
   # varName
   changes_var_tbl <- changes_var
   changes_var_tbl[1, "varName_new"] <- "new1"
@@ -149,9 +214,13 @@ test_that("Changes to GADSdat if tibble, varNames", {
   expect_equal(g1$labels$varName, c(rep("new1", 3), rep("VAR2", 2), rep("VAR3", 2)))
   expect_equal(names(g1$dat), c("new1", "VAR2", "VAR3"))
 
+  changes_var_df <- as.data.frame(changes_var_tbl)
+  g3 <- applyChangeMeta(changes_var_df, dfSAV)
+  expect_equal(g3$labels[, -1], dfSAV$labels[, -1])
+
 })
 
-test_that("Changes to GADSdat if tibble, value labels", {
+test_that("Changes to GADSdat if tibble or data.frame, value labels", {
   changes_val_tbl <- changes_val
   changes_val_tbl[1, "valLabel_new"] <- "new_miss"
   changes_val_tbl[2, "valLabel_new"] <- "new_miss2"
@@ -161,4 +230,17 @@ test_that("Changes to GADSdat if tibble, value labels", {
   expect_equal(g2$labels[, -7], dfSAV$labels[, -7])
   expect_equal(g2$labels$valLabel, c("new_miss", "new_miss2", "One", "missing", NA, "missing", NA))
   expect_equal(names(g2$dat), names(dfSAV$dat))
+
+  changes_val_df <- as.data.frame(changes_val_tbl)
+  g4 <- applyChangeMeta(changes_val_df, dfSAV)
+  expect_equal(g4$labels[, -7], dfSAV$labels[, -7])
 })
+
+
+test_that("sort value labels", {
+  unsorted <- dfSAV$labels[c(3:1), ]
+  out <- sort_value_labels(unsorted)
+
+  expect_equal(out, dfSAV$labels[1:3,])
+})
+