+++ tests/testthat/test_getChangeMeta.R
@@ -30,7 +30,8 @@ test_that("Extract value level meta change table", {
   changes_val3$missings_new <- "test"
   changes_val$value_new <- "test"
   changes_val4$value_new[1] <- NA
-  changes_val4$missings_new[1] <- "miss"
+  changes_val4[8, ] <- changes_val4[1, ]
+  changes_val4[8, c("value", "valLabel", "missings", "missings_new")] <- c(rep(NA, 3), "miss")
 
   expect_error(check_valChanges(changes_val2), "Irregular column names in changeTable.")
   expect_error(check_valChanges(changes_val3), "Irregular values in 'missings_new' column.")
@@ -49,16 +50,16 @@ test_that("Extract list of meta change tables for all_GADSdat", {
 })
 
 test_that("Check changeTable function", {
-  changes_var2 <- changes_var1 <- changes_var
+  changes_var3 <- changes_var2 <- changes_var1 <- changes_var
   dfSAV2 <- dfSAV
   changes_var1[1, "varLabel"] <- "sth"
   changes_val1 <- changes_val
   changes_val1[1, "varName"] <- "sth"
 
   expect_error(check_changeTable(dfSAV, changes_var1),
-               "GADSdat and changeTable are not compatible. Columns without '_new' should not be changed in the changeTable.")
+               "GADSdat and changeTable are not compatible in column 'varLabel'. Columns without '_new' should not be changed in the changeTable.")
   expect_error(check_changeTable(dfSAV, changes_val1),
-               "GADSdat and changeTable are not compatible. Columns without '_new' should not be changed in the changeTable.")
+               "GADSdat and changeTable are not compatible in column 'varName'. Columns without '_new' should not be changed in the changeTable.")
   expect_silent(check_changeTable(dfSAV, changes_var))
   expect_silent(check_changeTable(dfSAV, changes_val))
   expect_silent(check_changeTable(df1, getChangeMeta(df1, "value")))
@@ -66,4 +67,7 @@ test_that("Check changeTable function", {
   changes_var2[, "display_width"] <- 0L
   dfSAV2$labels$display_width <- 0
   expect_silent(check_changeTable(dfSAV2, changes_var2))
+
+  changes_var3[, "display_width"] <- NA_character_
+  expect_silent(check_changeTable(dfSAV2, changes_var3))
 })